+++ app/__init__.py
@@ -19,6 +19,10 @@ moment = Moment(app)
 
 from app.errors import bp as errors_bp
 app.register_blueprint(errors_bp)
+
+from app.auth import bp as auth_bp
+app.register_blueprint(auth_bp)
+
 from app import routes, models
 import logging
 import os
