+++ app/templates/auth/add_patient.html
