+++ app/templates/auth/email_patient.html
