+++ app/templates/auth/patient_info.html
@@ -39,7 +39,7 @@
               </div>
             </div>
              <div class="panel-footer">
-                    <a data-original-title="Email" data-toggle="tooltip" type="button" class="btn btn-sm btn-primary" href="{{ url_for('email_patient', email=patient.email) }}"><i class="glyphicon glyphicon-envelope"></i></a>
+                    <a data-original-title="Email" data-toggle="tooltip" type="button" class="btn btn-sm btn-primary" href="{{ url_for('auth.email_patient', email=patient.email) }}"><i class="glyphicon glyphicon-envelope"></i></a>
                     <span class="pull-right">
                         <a href="edit.html" data-original-title="Edit patient" data-toggle="tooltip" type="button" class="btn btn-sm btn-warning"><i class="glyphicon glyphicon-edit"></i></a>
                         <a data-original-title="Delete" data-toggle="tooltip" type="button" class="btn btn-sm btn-danger"><i class="glyphicon glyphicon-remove"></i></a>
