+++ app/templates/base.html
@@ -22,10 +22,10 @@
         </ul>
         <ul class="nav navbar-nav navbar-right">
           {% if current_user.is_anonymous %}
-          <li><a href="{{ url_for('login') }}">Login</a></li>
+          <li><a href="{{ url_for('auth.login') }}">Login</a></li>
           {% else %}
-          <li><a href="{{ url_for('all_patients', email=current_user.email) }}">Patients</a></li>
-          <li><a href="{{ url_for('logout') }}">Logout</a></li>
+          <li><a href="{{ url_for('auth.all_patients', email=current_user.email) }}">Patients</a></li>
+          <li><a href="{{ url_for('auth.logout') }}">Logout</a></li>
           {% endif %}
         </ul>
       </div>
