+++ app/templates/index.html
@@ -2,7 +2,7 @@
 
 {% block app_content %}
   {% if current_user.is_anonymous %}
-  <h1>Hello! <a href="{{ url_for('login') }} ">Login</a> to view your patients.</h1>
+  <h1>Hello! <a href="{{ url_for('auth.login') }} ">Login</a> to view your patients.</h1>
   {% else %}
   <h1>Hello {{ current_user.name }}</h1>
   {% endif %}
